+++ CHANGELOG.md
@@ -3,7 +3,9 @@
 ## 2020-03-02
 * Add `NHSD-Session-URID` header to specification.
 * Rename `from_asid` header to `NHSD-ASID`
+* New PDS sandbox search scenarios
 * Updating `Name` prefixes and suffixes to be an array of string, not string
+
 
 ## 2020-02-26
 * Add a config for dependabot so that security updates are automatically merged